Proper organization of the code
We had kind of mixed sources and distributables together, when we
started writing proper setup.py, this mess became apparent. What we want
to be able to do is the following, when we distribute 'sources' - we
want them to be able to be compiled on the machine of user's choice (we
want to do that better as well as a part of building those on a user's
machine, but let's ignore that for now.) But a runtime should only
require 'extension libraries'. That definition is not yet 'fully
realized' but we are getting there.

Our examples eg. `tshark.py` should make use of only the distributed
libraries and should not bother about the 'sources' being present. Right
now we are able to achieve that by `setup.py develop`, well almost.