`wsutil` common between wireshark2 and wireshark3
There are no differences in wsutil between wireshark2 and wireshark3
except for sone new Macros in `nstime.h`. Since we are not going to be
using most of the Macros, this is not a big problem, so abstracted it
out as common. Later if we find files diverging, they will go inside
their own `wsutil` submodule. For now they are common. As of now we do
not plan to make a wsutil library wrapper. If we decide to take that
path, they'll go into their individual submodules as well.