Getting started with API
This is not really an API that can be used, but a quick example wrapper
over a functionality that hides out details of which Wireshark version
is used underneath it. More thought needs to be given to what API might
look like and this is where next dev will happen. Right now we have only
one class `WishpyDissector` (an example really).