Making sure that there is enough space to copy Packet
In both v2 and v3 APIs, sometimes we receive big packets. In v3 APIs
we were simply ignoring stuff past the last and consequently didn't
face any issue, but in v2 API we are indeed facing issue and getting a
sigsegv. Now we allocate additional space if the packet is bigger than
some size (up to 4 times, done so that we do it as full pages)