Fixes `value_to_str` to make it faster
Running profiling on the code showed that `value_to_str` was the second
most expensive function in terms of it's own time. This was perhaps due
to the fact that we were building list as local variables (through a
referenced `epan_lib` object. Moving that out to the `class` level and
just making a local reference in the function appears to make the
function between 25-30% faster in terms of `tottime`, surely worth it.