Dereferencing `epan_lib` and `epan_ffi` objects to Class
In a number of functions, we were dereferencing in the fast path objects
from `epan_lib` and `epan_ffi`. This caused a bit of a performance hit
in those functions as these reference in large number of calls do make
a difference in terms of performance. Now we are keeping these values
handy at a class level. This has helped us get about 5% improvement in
performance.