Using wireshark's own `fvalue_to_string_repr`
Somehow we had missed this function and were doing everything by hand.
Instead we are now using wireshark's own function that gives us all
dissection capabilities without having to worry about how to display
values to user's. Wondering how I missed this? We still have to do some
work with regards to `quoting` and `unquoting` which wireshark should
ideally do for us as it has done all the hard work. But that's fine for
now.