Optinally 'serialize' header and data in Capturer
Right now we are simply putting, hdr and data received from the
interface on the queue as it is. These are 'pointers' and hence work
only if our 'dissector' is in same process. Converting it to sensible
Pythonic 'pickleable' structures will make it possible to move
capturer and dissector into separate processes. This would not work at
the moment.

This 'feature' was deemed important when we were trying to write unit
tests for the `WishpyDissectorQueuePython` where we realized, it's quite
hard without actually using it with our `capturer`. Also, this would
have been required at some point of time sooner to really make
dissectors fan out, just in case! :-)

Default value of serialize is `False` now, but will be perhaps later
changed to `True` and we might even give a default `json` or `pickle`
serializer (and their deserializer).