Added new `WishpyCapturerFileToQueue` and code refactoring
This is a pretty big change to the Capturer API.
- `LibpcapCapturer` is now renamed to `WishpyCapturerIfaceToQueue`.
- Added new `WishpyCapturerFileToQueue`
- Both Capturers derive from `WishpyCapturerQueue` which in turn
  derives from `WishpyCapturer` class.
- `serialize` option of 'start' is default to True now.

We discovered and fixed an interesting bug where we were using pointers
returned to us by `libpcap` in dissectors. This was fine as long as
dissector was picking packets and finishing it's work before the next
one arrived. This was a race condition. Unfortunately we didn't
encounter the race condition while capturing from the interface live
ever, may be because - between the time `libpcap` made new data
available, we were done dissecting. But this was easily exposed while
reading the PCAP File. So safe and correct way is - to SERDES the data
to/from the queue and hence `serialize` default is now `True`. Right now
cannot think of a 'safe' way of not doing this, with `serialize` to be
False.