Changes in dissector and example/tcpdump to accomodate new capturer
Changes to the example code to accomodate new capturer (Changed Class
names).

In the dissector, made a small change in `dissect_one_packet` to
explicitly break upon receiving a hdr that says 'stop' from the `run`
generator. This was observed with Pcap File - Even after the PCAP file
was completely dissected, we were still 'waiting' in the `run`
generator.