Making sure that the documentation builds on RTD
RTD tries to import our modules with the `sys.path` we have specified
in the `conf.py`. Since we have our own extension module, that is
'relative imported' and is not likely to be present there during the
build, we have to do just 'enough' to make sure RTD is happy. This means
giving a `fake_ext` with required names that will make the module
`import`able. We've to do some minor jugglary in the `capturer` module
as well. All of this finally works now so our API `autodoc` builds.