Fixes for better documentation
Moved `README.md` to `README.rst` so that it can be included inside
`docs/intro.rst` and we do not have to duplicate the effort. All the
required changes in the `README.rst` to support the RST format.

Fixed reference to `README.md` in `setup.py` and removed
`long_description_content_type`.

Test Added `_static` js file for making targets open in new tab (not
working right now but is not super important!)

Moved README-windows.md README-windows.rst