Updates to support Protocol Tree details
This support is right now early, but is broadly working. Following are
the updates -

Changed the calling convention of `print_json`, this is to keep
consistent with the way we want to call for `print_details`. Also it's
better readability wise.

We are passing `level` parameter to `print_json` without `pretty`. This
is done to be able to add `proto_tree` in the packet dissection json.

Added one more 'flag' `--add-proto-tree` in the `tshark.py` utility this
adds `proto_tree` attribute to packet json object.

Implementation of `print_dissected_tree_details` and
`print_dissected_tree_details_node` functions (for getting just the
Protocol Tree.)

Right now we do not have a clean way of calling `proto-tree-only`. We'll
fix that soon.