Elastic compliant json
In the Json generated there were a few 'duplicate' keys notably for
things that are wireshark 'filters' like `ip.addr`. While, Json parsers
are happy with it, elastic does not like it. Also, all the fields that
have a "." in the `abbrev` end up confusing elastic. Those `.`s are
converted to `_` s. Some more flattening is required too.

Also updated the 'scripts' to honour an elasticky flag, as this flag
slows down json dissection considerably especially for pypy! Almost all
the advantage is lost. Will need to look into this.